+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/table-menu/o-table-export-configuration.class.ts
@@ -1,13 +1,6 @@
 export class OTableExportConfiguration {
-  columns: Array<any>;
-  columnNames: object;
-  sqlTypes: object;
   service: string;
   serviceType: string;
-  data?: any[];
-  filter?: object;
-  mode: string;
-  entity: string;
   visibleButtons: string;
   options?: any;
 }