+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/table-menu/o-table-menu.component.ts
@@ -21,14 +21,13 @@ import { OTableMenu } from '../../../../../interfaces/o-table-menu.interface';
 import { IReportService } from '../../../../../interfaces/report-on-demand-service.interface';
 import { DialogService } from '../../../../../services/dialog.service';
 import { O_REPORT_SERVICE } from '../../../../../services/factories';
+import { OntimizeExportDataProviderService } from '../../../../../services/ontimize-export-data-provider.service';
 import { SnackBarService } from '../../../../../services/snackbar.service';
 import { OTranslateService } from '../../../../../services/translate/o-translate.service';
 import { OPermissions } from '../../../../../types/o-permissions.type';
 import { OTableMenuPermissions } from '../../../../../types/table/o-table-menu-permissions.type';
-import { Codes } from '../../../../../util/codes';
 import { PermissionsUtils } from '../../../../../util/permissions';
 import { Util } from '../../../../../util/util';
-import { OTableCellRendererImageComponent } from '../../../column/cell-renderer/image/o-table-cell-renderer-image.component';
 import { OColumn } from '../../../column/o-column.class';
 import { OTableComponent } from '../../../o-table.component';
 import { OTableGroupByColumnsDialogComponent } from '../../dialog';
@@ -43,7 +42,6 @@ import {
 import { OTableStoreFilterDialogComponent } from '../../dialog/store-filter/o-table-store-filter-dialog.component';
 import { OTableVisibleColumnsDialogComponent } from '../../dialog/visible-columns/o-table-visible-columns-dialog.component';
 import { OTableOptionComponent } from '../table-option/o-table-option.component';
-import { OTableExportConfiguration } from './o-table-export-configuration.class';
 
 
 
@@ -66,8 +64,10 @@ export const DEFAULT_INPUTS_O_TABLE_MENU = [
 
   // show-group-by-option [yes|no|true|false]: show group by menu option in the header menu
   'showGroupByOption: show-group-by-option',
+
   // show-reset-width-option [yes|no|true|false]: show reset width menu option in the header menu
   'showResetWidthOption: show-reset-width-option',
+
   // show-report-on-demand-option [yes|no|true|false]: show report on demand option in the header menu
   'showReportOnDemandOption: show-report-on-demand-option'
 ];
@@ -136,6 +136,7 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
 
   protected permissions: OTableMenuPermissions;
   protected mutationObservers: MutationObserver[] = [];
+  protected exportDataProvider: OntimizeExportDataProviderService;
 
 
   constructor(
@@ -148,6 +149,7 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
     this.dialogService = this.injector.get(DialogService);
     this.translateService = this.injector.get(OTranslateService);
     this.snackBarService = this.injector.get(SnackBarService);
+    this.exportDataProvider = this.injector.get(OntimizeExportDataProviderService)
   }
 
   ngOnInit() {
@@ -315,50 +317,15 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
   }
 
   onExportButtonClicked() {
-    const tableOptions = this.table.oTableOptions;
-    const exportCnfg: OTableExportConfiguration = new OTableExportConfiguration();
-
-    // get column's attr whose renderer is OTableCellRendererImageComponent
-    const colsNotIncluded: string[] = tableOptions.columns.filter(c => void 0 !== c.renderer && c.renderer instanceof OTableCellRendererImageComponent).map(c => c.attr);
-    colsNotIncluded.push(Codes.NAME_COLUMN_SELECT);
-    colsNotIncluded.push(Codes.NAME_COLUMN_EXPANDABLE);
-
-    // Table data/filters
-    switch (this.table.exportMode) {
-      case Codes.EXPORT_MODE_ALL:
-        exportCnfg.filter = this.table.getComponentFilter();
-        break;
-      case Codes.EXPORT_MODE_LOCAL:
-        exportCnfg.data = this.table.getAllRenderedValues();
-        colsNotIncluded.forEach(attr => exportCnfg.data.forEach(row => delete row[attr]));
-        break;
-      default:
-        exportCnfg.data = this.table.getRenderedValue();
-        colsNotIncluded.forEach(attr => exportCnfg.data.forEach(row => delete row[attr]));
-        break;
-    }
-    exportCnfg.mode = this.table.exportMode;
-    exportCnfg.entity = this.table.entity;
-
-    // Table columns
-    exportCnfg.columns = tableOptions.visibleColumns.filter(c => colsNotIncluded.indexOf(c) === -1);
-    // Table column names
-    const tableColumnNames = {};
-    tableOptions.visibleColumns.filter(c => colsNotIncluded.indexOf(c) === -1).forEach(c => {
-      const oColumn = tableOptions.columns.find(oc => oc.attr === c);
-      tableColumnNames[c] = this.translateService.get(oColumn.title ? oColumn.title : oColumn.attr);
-    });
-    exportCnfg.columnNames = tableColumnNames;
-    // Table column sqlTypes
-    exportCnfg.sqlTypes = this.table.getSqlTypes();
-    // Table service, needed for configuring ontimize export service with table service configuration
-    exportCnfg.service = this.table.service;
-    exportCnfg.serviceType = this.table.exportServiceType;
-    exportCnfg.visibleButtons = this.table.visibleExportDialogButtons;
-    exportCnfg.options = this.table.exportOptsTemplate;
+    this.exportDataProvider.initializeProvider(this.table);
 
     this.dialog.open(OTableExportDialogComponent, {
-      data: exportCnfg,
+      data: {
+        visibleButtons: this.table.visibleExportDialogButtons,
+        service: this.table.service,
+        serviceType: this.table.exportServiceType,
+        options: this.table.exportOptsTemplate
+      },
       disableClose: true,
       panelClass: ['o-dialog-class', 'o-table-dialog']
     });