+++ projects/ontimize-web-ngx/src/lib/config/o-providers.ts
@@ -6,7 +6,7 @@ import { combineLatest } from 'rxjs';
 
 import { AppConfig, O_INPUTS_OPTIONS } from '../config/app-config';
 import { appConfigFactory } from '../services/app-config.provider';
-import { ComponentStateServiceProvider, OntimizeAuthServiceProvider,  OntimizeServiceProvider, O_MAT_ERROR_OPTIONS } from '../services/factories';
+import { ComponentStateServiceProvider, ExportDataServiceProvider, OntimizeAuthServiceProvider, OntimizeExportServiceProvider, OntimizeServiceProvider, O_MAT_ERROR_OPTIONS } from '../services/factories';
 import { LocalStorageService } from '../services/local-storage.service';
 import { NavigationService } from '../services/navigation.service';
 import { OntimizeMatIconRegistry } from '../services/ontimize-icon-registry.service';
@@ -80,6 +80,8 @@ export const ONTIMIZE_PROVIDERS: Provider[] = [
   OntimizeServiceResponseAdapter,
   OntimizeAuthServiceProvider,
   ComponentStateServiceProvider,
+  ExportDataServiceProvider,
+  OntimizeExportServiceProvider,
   // disabled global ripple
   { provide: MAT_RIPPLE_GLOBAL_OPTIONS, useValue: { disabled: true } },
   { provide: O_MAT_ERROR_OPTIONS, useValue: {} },