+++ projects/ontimize-web-ngx/src/lib/util/codes.ts
@@ -118,6 +118,8 @@ export class Codes {
 
   public static APP_LAYOUT_MODE_DESKTOP: OAppLayoutMode = 'desktop';
   public static APP_LAYOUT_MODE_MOBILE: OAppLayoutMode = 'mobile';
+  public static VISIBLE_EXPORT_BUTTONS: string[] = ['xlsx', 'pdf', 'html'];
+  public static VISIBLE_EXPORT_BUTTONS3X: string[] = ['xlsx', 'pdf', 'csv'];
 
   public static DEFAULT_CHANGE_EVENT: CHANGE_EVENTS = 'onValueChange';
   public static TYPES_DATE_GROUPS = ["YEAR", "MONTH", "YEAR_MONTH", "YEAR_MONTH_DAY"];