+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * **o-grid,o-list**: Moved `quick-filter-appearance` to AbstractOServiceComponent ([0bc6c49](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/0bc6c49))
 * **o-form**: new `message-service-type` input that allows to override the default `o-form` CRUD operations success/error messages ([a490775e](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/a490775e)) Closes [#954](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/954). User can define this provider in the same way as in the `service-type` input.
 * **o-table, o-grid, o-list**: Added directive `o-table-toolbar`, `o-list-toolbar`, `o-grid-toolbar` to add custom content in position start or end of the toolbars ([1778a61](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/1778a61)) Closes [#1044](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1044)
+* New injection token `O_EXPORT_DATA_SERVICE` to allow customize the data and styles of the table exports ([ed16cd5](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/1778a61)) Closes [#993](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/993)
+* New parameter`exportConfiguration` in `app.config.ts`, export configuration object required only with only for Ontimize Boot version 3.9.0 or above ([442512c](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/442512c))
+
 
 ### Bug fixes
 * **o-real-input**: New attribute `strict` ([e466768f](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/e466768f)) ([e264a54](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/e264a54)) Closes [#1022](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1022)
@@ -18,6 +21,7 @@
   * Fixed bug in `AbstractOServiceBaseComponent` when searching for a column of type service render ([47a60682](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/47a60682)) Closes[#1010](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1010)
   * `o-table-columns-filter`: active by default. However, you can configure it unavailable in table header with `filter-column-active-by-default= 'no'`. ([6ceaefe5](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/6ceaefe5)) Closes [#1007](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1007)
   * Fix bug that filtering a value in the quick filter launchs the query to the backend when `pageable=no` ([367f3d7](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/367f3d7)) Closes [#1011](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1011)
+  * Fixed compatibility of Ontimize Boot 3.x.x exports ([7d741f8](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/7d741f8))
 * **o-email-input**: Fixed bug that incorrectly validated email inputs ([46c6e78](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/46c6e78)) Closes [#999](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/999)
 * **mat-error**: Bug [#1013](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1013) fixed. OntimizeWeb form component `mat-error` definition no longer interferes the Angular Material default definition.
   *NOTES*:
@@ -26,6 +30,8 @@
 * **service-type**: Fixes bug when multiple components use the same service-type ([380fda3](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/380fda3)) Closes [#1015](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1015)
 * **o-hour-input**: Fixing wrong `value-type="timestamp"` value saving ([f3d6f39c](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/f3d6f39c)) ([5b9abf17](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/5b9abf17)) Closes [#960](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/960)
 * **o-grid, o-table, o-list**: Fixes for the arrow button inside the quick filter being too far away from the magnifying glass button ([cdaf3db](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/cdaf3db)) Closes [#1056](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1056)
+
+
 
 ### BREAKING CHANGES
 * Due to the fix of the issue [#1015](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1015), to set the `service-type` attribute on the component, the provider defined in the module cannot set `useFactory`, instead use `useValue` because Ontimize Web is responsible for creating each instance for this service.
@@ -40,16 +46,33 @@
     useValue: StarWarsService
   }]
 })
+
 export class MyModule { }
 ```
 
+* `O_EXPORT_SERVICE` used to extend  `OntimizeExportService` but now extends  `OntimizeExportService` or `OntimizeExportService3X` depending on back server.
+
+| Ontimize server | Service |
+| ------- | ------- |
+| Ontimize EE, Ontimize Boot 2.X.X or lower version | `OntimizeExportService` |
+| Ontimize Boot 3.9.0 or higher |  `OntimizeExportService3X` |
+
+* Now the `configureService` method has only one `config` parameter and the `exportData` has only one parameter `format` in the `IExportService` interface since exportData is available through the provider defined with injection token `O_EXPORT_DATA_SERVICE`, check this [link](https://ontimizeweb.github.io/docs/v8/components/table/export) for more information ([2eaa9c7](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/2eaa9c7)) ([591d3c6](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/591d3c6))
+`
+// before
+  configureService(config: any, modeAll?: boolean): void;
+  exportData(data: any, format: string, entity?: string): Observable<any>;
+
+// after
+configureService(config: any): void;
+exportData(format: string): Observable<any>;
+`
 * Due to the improvement of the issue [#1044](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1044) the `o-table`, `o-list`, `o-grid` toolbars have been factored and this causes the following changes.
   * The css selectors related to the mat-toolbar will stop working in `o-table`, `o-list`, `o-grid` components.
   * New CSS class `o-table-toolbar`, `o-list-toolbar`, `o-grid-toolbar`
   * Unified the styles of the titles of the `o-table`, `o-list`, `o-grid` with new CSS class to `title` and therefore removed CSS class `table-title` and `o-list-title`
 
 * **Application configuration file `app.config,ts`**
-
 The default value of `serviceType` has been changed to `OntimizeEE` and the default value of `permissionsServiceType` has been changed to `OntimizeEEPermissions`([25cc5c8](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/25cc5c8))  Closes [#967](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/967)